Split graphql Schema up
