+++ src/server/graphql/schema.js
@@ -1,112 +1,14 @@
 import {
-  GraphQLBoolean,
-  GraphQLID,
-  GraphQLInt,
-  GraphQLList,
-  GraphQLNonNull,
-  GraphQLObjectType,
-  GraphQLSchema,
-  GraphQLString,
+  GraphQLObjectType, GraphQLSchema,
 } from 'graphql'
 
 import {
-  connectionArgs,
-  connectionDefinitions,
-  connectionFromArray,
-  cursorForObjectInConnection,
-  fromGlobalId,
-  globalIdField,
-  mutationWithClientMutationId,
-  nodeDefinitions,
-  toGlobalId,
-} from 'graphql-relay'
-
-import {
-  Todo,
-  User,
-  addTodo,
-  changeTodoStatus,
-  getTodo,
-  getTodos,
-  getUser,
   getViewer,
-  markAllTodos,
-  removeCompletedTodos,
-  removeTodo,
-  renameTodo,
 } from './database'
 
-const { nodeInterface, nodeField } = nodeDefinitions(
-  (globalId) => {
-    const { type, id } = fromGlobalId(globalId)
-    if (type === 'Todo') {
-      return getTodo(id)
-    } else if (type === 'User') {
-      return getUser(id)
-    }
-    return null
-  },
-  (obj) => {
-    if (obj instanceof Todo) {
-      return GraphQLTodo // eslint-disable-line no-use-before-define
-    } else if (obj instanceof User) {
-      return GraphQLUser // eslint-disable-line no-use-before-define
-    }
-    return null
-  }
-)
-
-const GraphQLTodo = new GraphQLObjectType({
-  name: 'Todo',
-  fields: {
-    id: globalIdField('Todo'),
-    text: {
-      type: GraphQLString,
-      resolve: (obj) => obj.text,
-    },
-    complete: {
-      type: GraphQLBoolean,
-      resolve: (obj) => obj.complete,
-    },
-  },
-  interfaces: [nodeInterface],
-})
-
-const {
-  connectionType: TodosConnection,
-  edgeType: GraphQLTodoEdge,
-} = connectionDefinitions({
-  name: 'Todo',
-  nodeType: GraphQLTodo,
-})
-
-const GraphQLUser = new GraphQLObjectType({
-  name: 'User',
-  fields: {
-    id: globalIdField('User'),
-    todos: {
-      type: TodosConnection,
-      args: {
-        status: {
-          type: GraphQLString,
-          defaultValue: 'any',
-        },
-        ...connectionArgs,
-      },
-      resolve: (obj, { status, ...args }) =>
-        connectionFromArray(getTodos(status), args),
-    },
-    totalCount: {
-      type: GraphQLInt,
-      resolve: () => getTodos().length,
-    },
-    completedCount: {
-      type: GraphQLInt,
-      resolve: () => getTodos('completed').length,
-    },
-  },
-  interfaces: [nodeInterface],
-})
+import { nodeField } from './interfaces'
+import * as todoMutations from './entities/Todo/mutations'
+import { GraphQLUser } from './entities/User'
 
 const Root = new GraphQLObjectType({
   name: 'Root',
@@ -119,147 +21,15 @@ const Root = new GraphQLObjectType({
   },
 })
 
-const GraphQLAddTodoMutation = mutationWithClientMutationId({
-  name: 'AddTodo',
-  inputFields: {
-    text: { type: new GraphQLNonNull(GraphQLString) },
-  },
-  outputFields: {
-    todoEdge: {
-      type: GraphQLTodoEdge,
-      resolve: ({ localTodoId }) => {
-        const todo = getTodo(localTodoId)
-        return {
-          cursor: cursorForObjectInConnection(getTodos(), todo),
-          node: todo,
-        }
-      },
-    },
-    viewer: {
-      type: GraphQLUser,
-      resolve: () => getViewer(),
-    },
-  },
-  mutateAndGetPayload: ({ text }) => {
-    const localTodoId = addTodo(text)
-    return { localTodoId }
-  },
-})
-
-const GraphQLChangeTodoStatusMutation = mutationWithClientMutationId({
-  name: 'ChangeTodoStatus',
-  inputFields: {
-    complete: { type: new GraphQLNonNull(GraphQLBoolean) },
-    id: { type: new GraphQLNonNull(GraphQLID) },
-  },
-  outputFields: {
-    todo: {
-      type: GraphQLTodo,
-      resolve: ({ localTodoId }) => getTodo(localTodoId),
-    },
-    viewer: {
-      type: GraphQLUser,
-      resolve: () => getViewer(),
-    },
-  },
-  mutateAndGetPayload: ({ id, complete }) => {
-    const localTodoId = fromGlobalId(id).id
-    changeTodoStatus(localTodoId, complete)
-    return { localTodoId }
-  },
-})
-
-const GraphQLMarkAllTodosMutation = mutationWithClientMutationId({
-  name: 'MarkAllTodos',
-  inputFields: {
-    complete: { type: new GraphQLNonNull(GraphQLBoolean) },
-  },
-  outputFields: {
-    changedTodos: {
-      type: new GraphQLList(GraphQLTodo),
-      resolve: ({ changedTodoLocalIds }) => changedTodoLocalIds.map(getTodo),
-    },
-    viewer: {
-      type: GraphQLUser,
-      resolve: () => getViewer(),
-    },
-  },
-  mutateAndGetPayload: ({ complete }) => {
-    const changedTodoLocalIds = markAllTodos(complete)
-    return { changedTodoLocalIds }
-  },
-})
-
-// TODO: Support plural deletes
-const GraphQLRemoveCompletedTodosMutation = mutationWithClientMutationId({
-  name: 'RemoveCompletedTodos',
-  outputFields: {
-    deletedTodoIds: {
-      type: new GraphQLList(GraphQLString),
-      resolve: ({ deletedTodoIds }) => deletedTodoIds,
-    },
-    viewer: {
-      type: GraphQLUser,
-      resolve: () => getViewer(),
-    },
-  },
-  mutateAndGetPayload: () => {
-    const deletedTodoLocalIds = removeCompletedTodos()
-    const deletedTodoIds = deletedTodoLocalIds.map(toGlobalId.bind(null, 'Todo'))
-    return { deletedTodoIds }
-  },
-})
-
-const GraphQLRemoveTodoMutation = mutationWithClientMutationId({
-  name: 'RemoveTodo',
-  inputFields: {
-    id: { type: new GraphQLNonNull(GraphQLID) },
-  },
-  outputFields: {
-    deletedTodoId: {
-      type: GraphQLID,
-      resolve: ({ id }) => id,
-    },
-    viewer: {
-      type: GraphQLUser,
-      resolve: () => getViewer(),
-    },
-  },
-  mutateAndGetPayload: ({ id }) => {
-    const localTodoId = fromGlobalId(id).id
-    removeTodo(localTodoId)
-    return { id }
-  },
-})
-
-const GraphQLRenameTodoMutation = mutationWithClientMutationId({
-  name: 'RenameTodo',
-  inputFields: {
-    id: { type: new GraphQLNonNull(GraphQLID) },
-    text: { type: new GraphQLNonNull(GraphQLString) },
-  },
-  outputFields: {
-    todo: {
-      type: GraphQLTodo,
-      resolve: ({ localTodoId }) => getTodo(localTodoId),
-    },
-  },
-  mutateAndGetPayload: ({ id, text }) => {
-    const localTodoId = fromGlobalId(id).id
-    renameTodo(localTodoId, text)
-    return { localTodoId }
-  },
-})
-
 const Mutation = new GraphQLObjectType({
   name: 'Mutation',
   fields: {
-    addTodo: GraphQLAddTodoMutation,
-    changeTodoStatus: GraphQLChangeTodoStatusMutation,
-    markAllTodos: GraphQLMarkAllTodosMutation,
-    removeCompletedTodos: GraphQLRemoveCompletedTodosMutation,
-    removeTodo: GraphQLRemoveTodoMutation,
-    renameTodo: GraphQLRenameTodoMutation,
+    addTodo: todoMutations.GraphQLAddTodoMutation,
+    changeTodoStatus: todoMutations.GraphQLChangeTodoStatusMutation,
+    markAllTodos: todoMutations.GraphQLMarkAllTodosMutation,
+    removeCompletedTodos: todoMutations.GraphQLRemoveCompletedTodosMutation,
+    removeTodo: todoMutations.GraphQLRemoveTodoMutation,
+    renameTodo: todoMutations.GraphQLRenameTodoMutation,
   },
 })
